Initial refactor work
