+++ neurocaps/analysis/__init__.py
@@ -1,4 +1,7 @@
-from .cap import CAP
+from .cap.cap import CAP
+
+# Easier import to clear cache
+from .cap._internals.spatial import build_tree, get_target_indices
 from .change_dtype import change_dtype
 from .merge import merge_dicts
 from .standardize import standardize
